@@ -4,6 +4,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@reduxjs/toolkit": "^2.8.2",
     "@testing-library/jest-dom": "^5.16.1",
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
@@ -13,16 +14,18 @@
     "@types/react-router-dom": "^5.3.3",
     "firebase": "^9.6.4",
     "rc-table": "^7.22.2",
-    "react": "^17.0.2",
+    "react": "^19.1.0",
     "react-csv": "^2.2.2",
     "react-custom-scrollbar": "^1.0.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^19.1.0",
+    "react-redux": "^9.2.0",
     "react-responsive": "^9.0.0-beta.6",
+    "react-router": "^7.6.2",
     "react-router-dom": "^6.2.1",
     "react-scripts": "^5.0.0",
     "react-select": "^5.8.0",
     "react-switch": "^7.0.0",
-    "react-table": "^7.7.0",
+    "react-table": "^7.8.0",
     "react-table-sticky": "^1.1.3",
     "styled-components": "^5.3.6",
     "typescript": "^4.6.2",
