@@ -1,35 +1,16 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import {createRoot} from 'react-dom/client';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 import FirebaseProvider from './components/FirebaseProvider';
-import {Routes, Route, HashRouter} from 'react-router-dom';
-
-ReactDOM.render(
+const root  = createRoot(document.getElementById('root')as HTMLElement);
+root.render(
   <React.StrictMode>
-    <HashRouter>
-      <Routes>
-        <Route
-          path="/women"
-          element={
-            <FirebaseProvider men={false}>
-              <App />
-            </FirebaseProvider>
-          }
-        ></Route>
-        <Route
-          path="*"
-          element={
-            <FirebaseProvider men>
-              <App />
-            </FirebaseProvider>
-          }
-        ></Route>
-      </Routes>
-    </HashRouter>
+    <FirebaseProvider>
+      <App />
+    </FirebaseProvider>
   </React.StrictMode>,
-  document.getElementById('root')
 );
 
 // If you want to start measuring performance in your app, pass a function
