@@ -1,7 +1,7 @@
-import { totalData } from '../types'
+import { yearlyTotals } from '../types'
 
 
-export const getLatestYear = (data:totalData) : string =>{
+export const getLatestYear = (data:yearlyTotals) : string =>{
   const keys = Object.keys(data);
   const [sorted] = keys.sort((a,b)=>{
     const yearA = +a.slice(0,4);
